♻️ Refactor: 폴더 구분
