+++ src/common/components/etc/BoxedFile/BoxedFile.client.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import clsx from 'clsx';
 
 import FeaturedFileIcons from '@icons/FeaturedFileIcons/FeaturedFileIcons';
