+++ src/common/components/etc/BoxedFile/BoxedFile.types.ts
