+++ src/common/components/inputs/InputInviteMember/InputInviteMember.client.tsx
@@ -1,4 +1,6 @@
-import React, { useEffect, useState } from 'react';
+'use client';
+
+import { useEffect, useState } from 'react';
 
 import clsx from 'clsx';
