+++ src/common/components/inputs/InputOnboarding/InputOnboarding.types.ts
@@ -13,7 +13,7 @@ export enum OnboardingMode {
   DEFAULT = 'DEFAULT',
   EDITABLE = 'EDITABLE',
 }
-export interface OnBoardingProps {
+export interface OnboardingProps {
   mode?: OnboardingMode;
   title: string;
   inputValue: string;
