+++ src/common/components/inputs/InputSearchBox/InputSearchBox.client.tsx
@@ -1,4 +1,4 @@
-import clsx from 'clsx';
+'use client';
 
 import IndividualIcons from '@common/components/icons/IndividualIcons/IndividualIcons';
 import { IndividualIconsState } from '@common/components/icons/IndividualIcons/IndividualIcons.types';
