+++ src/common/components/inputs/InputSearchBox/InputSearchBox.types.ts
