+++ src/common/components/select-box/SelectBoxOption/SelectBoxOption.client.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { useState } from 'react';
 
 import clsx from 'clsx';
