+++ src/common/components/select-box/SelectBoxOption/SelectBoxOption.types.ts
