+++ src/common/components/select-box/SelectBoxTag/SelectBoxTag.client.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import clsx from 'clsx';
 
 const SelectBoxTag = ({ onClick, isSelected = false, onToggle, label }: SelectBoxTagProps) => {
