+++ src/common/components/select-box/SelectBoxTag/SelectBoxTag.types.ts
